+++ test/models/serializable_test.rb
@@ -46,6 +46,10 @@ class SerializableTest < ActiveSupport::TestCase
     assert_key "username", @user.as_json({ only: :username, except: [:email].freeze }.freeze)["user"]
   end
 
+  test 'constant `BLACKLIST_FOR_SERIALIZATION` is deprecated' do
+    assert_deprecated { Devise::Models::Authenticatable::BLACKLIST_FOR_SERIALIZATION }
+  end
+
   def assert_key(key, subject)
     assert subject.key?(key), "Expected #{subject.inspect} to have key #{key.inspect}"
   end